not much

+++ mainboard/firmware/drivers.h
@@ -9,7 +9,7 @@ struct driver_packet_set {
     float targetSpeed;
     int16_t pwm;
     uint32_t padding1;
-    uint8_t padding2;
+    uint16_t padding2;
 } __attribute__((packed));
 
 struct driver_packet_ans {
